# Socify-Companion/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SOCIFY API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List workspaces for current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
    post:
      operationId: createWorkspace
      tags: [workspaces]
      summary: Create workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
  /workspaces/{workspaceId}:
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Get workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
    patch:
      operationId: updateWorkspace
      tags: [workspaces]
      summary: Update workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkspaceBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"

  # ─── SOCIAL ACCOUNTS ────────────────────────────────────────────────────────
  /social-accounts:
    get:
      operationId: listSocialAccounts
      tags: [social-accounts]
      summary: List social accounts
      parameters:
        - name: workspaceId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SocialAccount"
    post:
      operationId: createSocialAccount
      tags: [social-accounts]
      summary: Connect a social account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSocialAccountBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialAccount"
  /social-accounts/{accountId}:
    delete:
      operationId: deleteSocialAccount
      tags: [social-accounts]
      summary: Disconnect social account
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── POSTS ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Auth ──
...
    Workspace:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        logo:
          type: string
          nullable: true
        plan:
          type: string
          enum: [free, pro, agency, enterprise]
        membersCount:
          type: integer
        socialAccountsCount:
          type: integer
        postsThisMonth:
          type: integer
        createdAt:
          type: string
      required: [id, name, slug, plan, membersCount, socialAccountsCount, postsThisMonth, createdAt]
    CreateWorkspaceBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
      required: [name, slug]
    UpdateWorkspaceBody:
      type: object
      properties:
        name:
          type: string
        logo:
          type: string

    # ── SocialAccount ──
    SocialAccount:
      type: object
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        platform:
          type: string
          enum: [facebook, instagram, twitter, linkedin, tiktok, youtube, pinterest, threads]
        accountType:
          type: string
          enum: [page, profile, business]
        name:
          type: string
        handle:
          type: string
        avatar:
          type: string
          nullable: true
        isActive:
          type: boolean
        followers:
          type: integer
        following:
          type: integer
        avgEngagementRate:
          type: number
          nullable: true
        lastSyncAt:
          type: string
          nullable: true
        connectedAt:
          type: string
      required: [id, workspaceId, platform, accountType, name, handle, isActive, followers, following, connectedAt]
    CreateSocialAccountBody:
      type: object
      properties:
        workspaceId:
          type: integer
        platform:
          type: string
        accountType:
          type: string
        name:
          type: string
        handle:
          type: string
        followers:
          type: integer
      required: [workspaceId, platform, accountType, name, handle, followers]

    # ── Post ──
...